+++ android/app/src/main/kotlin/com/ccextractor/ultimate_alarm_clock/ultimate_alarm_clock/TimerBroadcasts/TimerFunctions.kt
@@ -1,4 +1,4 @@
-package com.example.ultimate_alarm_clocks
+package com.ccextractor.ultimate_alarm_clocks
 import android.content.ContentValues
 import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
