+++ android/app/src/main/kotlin/com/ccextractor/ultimate_alarm_clock/ultimate_alarm_clock/TimerBroadcasts/TimerNotification.kt
@@ -1,51 +1,34 @@
-package com.example.ultimate_alarm_clock
-
-import android.annotation.SuppressLint
-import android.app.AlarmManager
+package com.ccextractor.ultimate_alarm_clock
+import android.app.NotificationChannel
+import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.os.SystemClock
-import android.app.NotificationChannel
-import android.app.NotificationManager
 import android.os.Build
-import android.os.CountDownTimer
 import androidx.core.app.NotificationCompat
-import com.example.ultimate_alarm_clocks.getLatestTimer
+import com.ccextractor.ultimate_alarm_clocks.getLatestTimer
 
 
-class BootReceiver : BroadcastReceiver() {
-
+class TimerNotification : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
-
-        if (intent.action == Intent.ACTION_BOOT_COMPLETED) {
-
-            val dbHelper = DatabaseHelper(context)
-            val db = dbHelper.readableDatabase
-            val ringTime = getLatestAlarm(db, true)
-            db.close()
-            if (ringTime != null) {
-                scheduleAlarm(ringTime, context)
+        val timerdbhelper = TimerDatabaseHelper(context)
+        val timerdb = timerdbhelper.readableDatabase
+        val time = getLatestTimer(timerdb)
+        timerdb.close()
+        var notificationManager =
+            context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        val commonTimer = CommonTimerManager.getCommonTimer(object : TimerListener {
+            override fun onTick(millisUntilFinished: Long) {
+                showTimerNotification(millisUntilFinished, "Timer", context)
             }
 
-            val timerdbhelper = TimerDatabaseHelper(context)
-            val timerdb = timerdbhelper.readableDatabase
-            val time = getLatestTimer(timerdb)
-            timerdb.close()
-            var notificationManager =
-                context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            val commonTimer = CommonTimerManager.getCommonTimer(object : TimerListener {
-                override fun onTick(millisUntilFinished: Long) {
-                    println(millisUntilFinished)
-                    showTimerNotification(millisUntilFinished, "Timer", context)
+            override fun onFinish() {
+                notificationManager.cancel(1)
+            }
+        })
 
-                }
-
-                override fun onFinish() {
-                    notificationManager.cancel(1)
-                }
-            })
+        if (intent.action == "com.ccextractor.ultimate_alarm_clock.START_TIMERNOTIF" || intent.action == Intent.ACTION_BOOT_COMPLETED) {
             createNotificationChannel(context)
 
 
@@ -57,23 +40,14 @@ class BootReceiver : BroadcastReceiver() {
 
             }
 
+        }
+        if (intent.action == "com.ccextractor.ultimate_alarm_clock.STOP_TIMERNOTIF") {
+
+            commonTimer.stopTimer()
 
         }
-    }
 
-    fun scheduleAlarm(milliSeconds: Long, context: Context) {
-        val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-        val intent = Intent(context, AlarmReceiver::class.java)
-        val pendingIntent = PendingIntent.getBroadcast(
-            context,
-            1,
-            intent,
-            PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_MUTABLE
-        )
 
-        // Schedule the alarm
-        val triggerTime = SystemClock.elapsedRealtime() + milliSeconds
-        alarmManager.setExact(AlarmManager.ELAPSED_REALTIME_WAKEUP, triggerTime, pendingIntent)
     }
 
     private fun createNotificationChannel(context: Context) {
@@ -95,7 +69,7 @@ class BootReceiver : BroadcastReceiver() {
         var notificationManager =
             context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
         val deleteIntent = Intent(context, TimerNotification::class.java)
-        deleteIntent.action = "com.example.ultimate_alarm_clock.STOP_TIMERNOTIF"
+        deleteIntent.action = "com.ccextractor.ultimate_alarm_clock.STOP_TIMERNOTIF"
         val deletePendingIntent = PendingIntent.getBroadcast(
             context, 5, deleteIntent,
             PendingIntent.FLAG_IMMUTABLE
